implement display of notification for new chat msgs in footer menu component
