+++ app/src/main/java/io/actifit/fitnesstracker/actifitfitnesstracker/BaseActivity.java
@@ -1,5 +1,6 @@
 package io.actifit.fitnesstracker.actifitfitnesstracker;
 
+import android.animation.ValueAnimator;
 import android.app.AlertDialog;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -20,11 +21,26 @@ import static io.actifit.fitnesstracker.actifitfitnesstracker.MainActivity.tutVi
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.Locale;
+
 import androidx.annotation.LayoutRes;
+
+import com.android.volley.Request;
+import com.android.volley.RequestQueue;
+import com.android.volley.toolbox.JsonArrayRequest;
+import com.android.volley.toolbox.Volley;
+
+import org.json.JSONArray;
+import org.json.JSONObject;
 
 public class BaseActivity extends AppCompatActivity {
 
     private static final String TAG = "BaseActivity";
+    public static int lastChatCount = 0;
+
+    ValueAnimator valueAnimator;
+    TextView chatNotifCount;
+    RequestQueue queue;
 
     @Override
     public void setContentView(@LayoutRes int layoutResID) {
@@ -90,6 +106,7 @@ public class BaseActivity extends AppCompatActivity {
         TextView btnMarket = findViewById(R.id.btn_view_market);
         TextView btnStepHistory = findViewById(R.id.btn_view_history);
         TextView btnViewSocial = findViewById(R.id.btn_view_social);
+        chatNotifCount = findViewById(R.id.chat_notif_count);
 
 
         if (btnHome != null){
@@ -118,8 +135,8 @@ public class BaseActivity extends AppCompatActivity {
         if (btnChat != null) {
             btnChat.setOnClickListener(view -> {
                 // Call the methods now part of BaseActivity
-                storeNotifDate(new Date(), "chat"); // Added type for clarity
-                storeNotifCount(MainActivity.lastChatCount); // Use variable now part of BaseActivity
+                storeNotifDate(new Date(), ""); // Added type for clarity
+                storeNotifCount(lastChatCount); // Use variable now part of BaseActivity
 
                 // Open chat using 'this' as Context
                 ChatDialogFragment chatDialog = ChatDialogFragment.newInstance(this);
@@ -264,6 +281,16 @@ public class BaseActivity extends AppCompatActivity {
                 dialog.show(getSupportFragmentManager(), "video_upload_fragment");
             });
         }
+
+
+        new Thread(() -> {
+            runOnUiThread(() -> {
+                queue = Volley.newRequestQueue(this);
+                //load chat notifications
+                loadChatNotif(queue);
+            });
+        }).start();
+
         updateNavigationButtonStates(); // Highlight the current activity's button
     }
 
@@ -274,7 +301,7 @@ public class BaseActivity extends AppCompatActivity {
 
         //Date date = new Date();
         String strDate = "";
-        if (!dateStr.equals("")){
+        if (!dateStr.isEmpty()){
             strDate = dateStr;
         }else{
             DateFormat dateFormat = new SimpleDateFormat("yyyyMMdd");
@@ -282,16 +309,197 @@ public class BaseActivity extends AppCompatActivity {
         }
 
         //String strDate = dateFormat.format(date);
-
+        Log.d(TAG, "strDate:"+strDate);
         editor.putString(getString(R.string.sting_chat_update), strDate);
-        editor.commit();
+        editor.apply();
     }
 
     public void storeNotifCount(int commCount){
         SharedPreferences shPrefs = getSharedPreferences("actifitSets",MODE_PRIVATE);
         SharedPreferences.Editor editor = shPrefs.edit();
         editor.putInt(getString(R.string.sting_chat_comm_count), commCount);
-        editor.commit();
+        editor.apply();
+    }
+
+
+    //check community notifications
+    private void loadCommunityNotif(RequestQueue queue, String lastChatDate, int commChatCount){
+
+        String notificationsUrl = getString(R.string.sting_chat_query_comm);
+
+        // Request the transactions of the user first via JsonArrayRequest
+        // according to our data format
+        JsonArrayRequest transactionRequest = new JsonArrayRequest(Request.Method.GET,
+                notificationsUrl, null, notificationsListArray -> {
+            try {
+                JSONObject todayObj = null;
+                if (notificationsListArray.length() > 0) {
+                    //today's stats
+                    todayObj = notificationsListArray.getJSONArray(1).getJSONArray(0).getJSONObject(6);//dual array structure
+                    //we have new messages today
+                    if (todayObj.has(getString(R.string.actifit_comm))){
+                        //new messages today, notify user
+                        SimpleDateFormat outputFormat = new SimpleDateFormat("yyyyMMdd", Locale.getDefault());
+                        String todayDate = outputFormat.format(new Date());
+                        lastChatCount = todayObj.getInt(getString(R.string.actifit_comm));
+                        if ((Integer.parseInt(todayDate) > Integer.parseInt(lastChatDate))){
+                            //storeNotifDate(null, notifDate);
+                            renderChatData();
+                        }else if ((Integer.parseInt(todayDate) == Integer.parseInt(lastChatDate))
+                                && lastChatCount > commChatCount){
+                            storeNotifCount(lastChatCount);
+                            renderChatData();
+                        }
+                    }
+                }
+
+            }catch(Exception ex){
+                ex.printStackTrace();
+            }
+        }, error -> {
+            //hide dialog
+            error.printStackTrace();
+
+        });
+
+        // Add transaction request to be processed
+        queue.add(transactionRequest);
+    }
+
+
+    private void loadChatNotif(RequestQueue queue){
+        //start by hiding notifierh
+        hideChatData();
+
+
+
+        //for testing, set custom date
+        /*
+        storeNotifDate(null, "20230924");
+        storeNotifCount(0);
+
+         */
+        //grab last stored update date for sting chat notifications
+        SharedPreferences sharedPreferences = getSharedPreferences("actifitSets",MODE_PRIVATE);
+
+        /*
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        editor.remove(getString(R.string.sting_chat_update));
+        editor.apply();*/
+
+        String lastChatDate = sharedPreferences.getString(getString(R.string.sting_chat_update),"");
+        int commChatCount = sharedPreferences.getInt(getString(R.string.sting_chat_comm_count),0);
+        Log.d(TAG, "commChatCount:"+commChatCount);
+        Log.d(TAG, "lastChatDate:"+lastChatDate);
+        if (!lastChatDate.isEmpty()){//(false){//
+            //we have a stored date, grab it
+
+            // This holds the url to connect to the API and grab the transactions.
+            // We append to it the username
+            String notificationsUrl = getString(R.string.sting_chat_query_user)
+                    .replace("_USER_", MainActivity.username);
+
+            // Request the transactions of the user first via JsonArrayRequest
+            // according to our data format
+            JsonArrayRequest transactionRequest = new JsonArrayRequest(Request.Method.GET,
+                    notificationsUrl, null, notificationsListArray -> {
+
+                boolean foundNew = false;
+                // Handle the result
+                try {
+                    JSONArray innerArray = null;
+                    if (notificationsListArray.length() > 0  ){
+                        innerArray = notificationsListArray.getJSONArray(1);
+                    }
+                    for (int i = 0; i < innerArray.length(); i++) {
+                        try{
+                            // Retrieve each JSON object within the JSON array
+                            JSONObject jsonObject = innerArray.getJSONObject(i);
+
+                            //SimpleDateFormat inputFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSX", Locale.getDefault());
+                            SimpleDateFormat inputFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS", Locale.getDefault());
+                            Date date = inputFormat.parse(jsonObject.getString("date"));
+
+                            SimpleDateFormat outputFormat = new SimpleDateFormat("yyyyMMdd", Locale.getDefault());
+                            String notifDate = outputFormat.format(date);
+                            if ((Integer.parseInt(notifDate) > Integer.parseInt(lastChatDate))){
+                                //storeNotifDate(null, notifDate);
+                                renderChatData();
+                                foundNew = true;
+                                break;
+                            }
+                        }catch(Exception exc){
+                            exc.printStackTrace();
+                        }
+                    }
+                    if (!foundNew){
+                        //check community notifications
+                        loadCommunityNotif(queue, lastChatDate, commChatCount);
+                    }
+
+                    //actifitTransactions.setText("Response is: "+ response);
+                }catch (Exception e) {
+                    //hide dialog
+
+                    e.printStackTrace();
+                }
+            }, error -> {
+                //hide dialog
+                error.printStackTrace();
+
+            });
+
+
+            // Add transaction request to be processed
+            queue.add(transactionRequest);
+
+        }else{
+            //default case
+            renderChatData();
+        }
+
+    }
+
+    private void renderChatData(){
+        if (chatNotifCount == null) return;
+        chatNotifCount.setVisibility(View.VISIBLE);
+
+        // Create a new ValueAnimator object.
+        if (valueAnimator == null) {
+            valueAnimator = new ValueAnimator();
+
+            // Set the duration of the animation.
+            valueAnimator.setDuration(1000);
+            valueAnimator.setRepeatCount(ValueAnimator.INFINITE);
+            valueAnimator.setRepeatMode(ValueAnimator.REVERSE);
+
+            // Set the valueFrom and valueTo properties of the ValueAnimator object.
+            //valueAnimator.setFloatValues(Color.RED, Color.parseColor("#FFFFFF"), Color.RED);
+            valueAnimator.setFloatValues(getResources().getColor(R.color.actifitRed), getResources().getColor(R.color.colorWhite));//Color.parseColor("#FFFFFF"));
+
+
+            // Add an AnimatorUpdateListener to the ValueAnimator object.
+            valueAnimator.addUpdateListener(animator -> {
+                // Set the background color of the TextView object to the current value of the ValueAnimator object.
+                int animatedValue = (int) (float) animator.getAnimatedValue();
+
+                chatNotifCount.setTextColor(animatedValue);
+            });
+        }
+
+// Start the ValueAnimator object.
+        if (!valueAnimator.isRunning()) {
+            valueAnimator.start();
+        }
+
+    }
+
+    private void hideChatData(){
+        if (chatNotifCount==null) return;
+        chatNotifCount.setVisibility(View.GONE);
+        if (valueAnimator != null && valueAnimator.isRunning()) {
+            valueAnimator.cancel();
+        }
     }
 
     @Override
@@ -302,5 +510,11 @@ public class BaseActivity extends AppCompatActivity {
         if (findViewById(R.id.btn_view_social) != null) {
             updateNavigationButtonStates();
         }
+        new Thread(() -> {
+            runOnUiThread(() -> {
+                //load chat notifications
+                loadChatNotif(queue);
+            });
+        }).start();
     }
 }
